Add some post merge conflict fixes and resolved AllActivities component file
