+++ ui/src/components/app/App.js
@@ -6,24 +6,16 @@ import LandingPage from "../landingPage/LandingPage";
 import AllActivities from "../allActivities/AllActivities";
 import ToDos from "../todo/ToDos";
 import ActivityInfoView from "../activityInfoView/ActivityInfoView";
-import { Form } from 'react-router-dom';
-import activities from "../../api/dummyData";
-import { useEffect, useState } from 'react';
 import ErrorPage from "../errorPage/ErrorPage";
 
-
 function App() {
-	const [allActivities, setAllActivities] = useState([])
 
-	useEffect(() => {
-		setAllActivities(activities)
-	})
   return (
     <main className="main-app">
       <HeaderNav />
 			<Routes>
 				<Route path='/' element={<LandingPage />}/>
-				<Route path='/AllActivities' element={<AllActivities allActivities={allActivities}/>}/>
+				<Route path='/AllActivities' element={<AllActivities />}/>
 				<Route path='/ToDos' element={<ToDos />} />
 				<Route path='/ActivityInfoView' element={<ActivityInfoView />}/>
 				<Route path='*' element={<ErrorPage />}/>
@@ -42,7 +34,6 @@ function App() {
         <p className="light-blue">lightBlue&nbsp;</p>
         <p className="pink">pink&nbsp;</p>
       </section>
-
     </main>
   );
 }
