+++ ui/src/components/app/App.scss
@@ -1,7 +1,5 @@
 @use "src/index.scss" as *;
-.App {
-  text-align: center;
-}
+
 .main-app {
 	text-align: center;
   height: 100vh;
@@ -13,42 +11,6 @@ main {
 	display: flex;
 	flex-direction: column;
 	align-items: center;
-}
-
-.App-logo {
-	height: 40vmin;
-	pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-	.App-logo {
-		animation: App-logo-spin infinite 20s linear;
-	}
-}
-
-.App-header {
-	background-color: #282c34;
-	min-height: 100vh;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	font-size: calc(10px + 2vmin);
-	color: white;
-}
-
-.App-link {
-	color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-	from {
-		transform: rotate(0deg);
-	}
-
-	to {
-		transform: rotate(360deg);
-	}
 }
 
 .trying-out-colors {
