+++ ui/src/components/thumbnailActivity/ActivityCard.scss
@@ -1,4 +1,5 @@
 @use "src/index.scss" as *;
+
 .activity-card {
 	background-color: brown;
 	border-radius: 4px;
@@ -16,8 +17,4 @@ img {
 	color: black;
   text-decoration: none;
   margin: 0;
-}
-
-.title {
-	
 }
